+++ src/main/java/com/plummy/cyhunters/Camera/Camera.java
@@ -1,7 +1,7 @@
-package com.plummy.cyhunters.Assets;
+package com.plummy.cyhunters.Camera;
 
-import com.plummy.cyhunters.Assets.Interfaces.ICamera;
-import com.plummy.cyhunters.Assets.Interfaces.IGamePlayer;
+import com.plummy.cyhunters.Iterfaces.ICamera;
+import com.plummy.cyhunters.Iterfaces.IGamePlayer;
 import org.bukkit.Location;
 import org.bukkit.util.Vector;
