+++ src/main/java/com/plummy/cyhunters/Camera/CameraSelector.java
@@ -1,9 +1,9 @@
-package com.plummy.cyhunters.Assets;
+package com.plummy.cyhunters.Camera;
 
-import com.plummy.cyhunters.Assets.Enums.Role;
-import com.plummy.cyhunters.Assets.Interfaces.ICamera;
-import com.plummy.cyhunters.Assets.Interfaces.ICameraSelector;
-import com.plummy.cyhunters.Assets.Interfaces.IGamePlayer;
+import com.plummy.cyhunters.Enums.Role;
+import com.plummy.cyhunters.Iterfaces.ICamera;
+import com.plummy.cyhunters.Iterfaces.ICameraSelector;
+import com.plummy.cyhunters.Iterfaces.IGamePlayer;
 
 import java.util.Collections;
 import java.util.List;
