+++ src/main/java/com/plummy/cyhunters/CyHunters.java
@@ -1,11 +1,15 @@
 package com.plummy.cyhunters;
 
-import com.plummy.cyhunters.Assets.*;
-import com.plummy.cyhunters.Assets.Interfaces.IGame;
+import com.plummy.cyhunters.Iterfaces.IGame;
+import com.plummy.cyhunters.Camera.CameraManager;
 import com.plummy.cyhunters.Commands.CyHuntersCommand;
 import com.plummy.cyhunters.Commands.CyHuntersCompleter;
+import com.plummy.cyhunters.Game.Game;
+import com.plummy.cyhunters.Game.GameBoard;
+import com.plummy.cyhunters.Game.LocationFinder;
 import com.plummy.cyhunters.Listeners.CameraListener;
 import com.plummy.cyhunters.Listeners.PlayerListener;
+import com.plummy.cyhunters.Player.PlayerManager;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
