+++ src/main/java/com/plummy/cyhunters/Enums/PlayerState.java
@@ -1,4 +1,4 @@
-package com.plummy.cyhunters.Assets.Enums;
+package com.plummy.cyhunters.Enums;
 
 public enum PlayerState {
     PLAYING,
