+++ src/main/java/com/plummy/cyhunters/Enums/Role.java
@@ -1,4 +1,4 @@
-package com.plummy.cyhunters.Assets.Enums;
+package com.plummy.cyhunters.Enums;
 
 public enum Role {
     SPEEDRUNNER,
