+++ src/main/java/com/plummy/cyhunters/Iterfaces/IGameBoard.java
@@ -1,4 +1,4 @@
-package com.plummy.cyhunters.Assets.Interfaces;
+package com.plummy.cyhunters.Iterfaces;
 
 public interface IGameBoard {
     void updateBoard();
