+++ src/main/java/com/plummy/cyhunters/Iterfaces/IPlayerManager.java
@@ -1,4 +1,4 @@
-package com.plummy.cyhunters.Assets.Interfaces;
+package com.plummy.cyhunters.Iterfaces;
 
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
