+++ src/main/java/com/plummy/cyhunters/Player/PlayerManager.java
@@ -1,16 +1,14 @@
-package com.plummy.cyhunters.Assets;
+package com.plummy.cyhunters.Player;
 
-import com.plummy.cyhunters.Assets.Enums.PlayerState;
-import com.plummy.cyhunters.Assets.Enums.Role;
-import com.plummy.cyhunters.Assets.Interfaces.IGamePlayer;
-import com.plummy.cyhunters.Assets.Interfaces.IPlayerManager;
+import com.plummy.cyhunters.Enums.PlayerState;
+import com.plummy.cyhunters.Enums.Role;
+import com.plummy.cyhunters.Iterfaces.IGamePlayer;
+import com.plummy.cyhunters.Iterfaces.IPlayerManager;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 
 import java.util.*;
-
-import static com.plummy.cyhunters.CyHunters.logger;
 
 public class PlayerManager implements IPlayerManager {
     private final Map<UUID, IGamePlayer> players;
